Translate encode_output into CakeML
This is one of the functions that should be in the Botworld preamble.

One problem: the sexp type uses char lists, but we probably would prefer
to target CakeML strings. The only obvious solution is to create a copy
of sexp stuff over mlstrings instead. I have asked the CakeML dev list
for other ideas.